backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Stock Data
yfinance==0.2.33
pandas==2.1.3
numpy==1.26.2
pykrx==1.0.45
beautifulsoup4==4.12.3
lxml==5.2.1
# TA-Lib==0.4.28  # Temporarily disabled due to build issues

# Technical Indicators
# Using pure pandas/numpy implementation instead of pandas-ta

# ML/AI
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.17.0
statsmodels==0.14.1

# Task Queue
celery==5.3.4
redis==5.0.1

# Monitoring
prometheus-client==0.19.0
influxdb-client==1.38.0

# Utilities
httpx==0.25.2
schedule==1.2.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
locust==2.20.0
requests==2.31.0